3847 hooks reshuffling
